move get_svn_version() from fba_launch_io.py to utils.py
